ifrt_proxy: use int64_t for shape-derived string element count
Per review: the string element count is derived from a shape
(Shape::num_elements() returns int64_t), so use signed int64_t for the
count, its parameter, and the comparison, following the codebase
convention and the Google style guide, instead of size_t.

+++ xla/python/ifrt_proxy/client/array.cc
@@ -723,7 +723,7 @@ tsl::Future<> Array::CopyToStringHostBuffer(
                    host_buffer_store = rpc_helper_->host_buffer_store(),
                    host_buffer_handle,
                    dst_buffer = static_cast<absl::Cord*>(data),
-                   num_elements = static_cast<size_t>(shape_.num_elements())](
+                   num_elements = shape_.num_elements()](
                       absl::StatusOr<std::shared_ptr<CopyToHostBufferResponse>>
                           resp) mutable {
     if (!resp.ok()) {
